refactor(server): refactor models
