--- server/src/lib/redis/model.ts
@@ -1,10 +1,0 @@
-export interface SlugInput {
-	url: string;
-	ext: string;
-	expiresAt: number;
-}
-
-export interface SlugOutput {
-	url: string;
-	ext: string;
-}
