+++ server/src/routes/files/controllers/upload/model.ts
@@ -1,4 +1,11 @@
-import { SlugInput, SlugOutput } from "lib/redis/model";
+import {
+	AckInput,
+	AckOutput,
+	FileStorage,
+	SlugStorage,
+	UploadInput,
+	UploadOutput,
+} from "model/model";
 
 export type CreateUploadController = (deps: {
 	fileStorage: FileStorage;
@@ -10,42 +17,4 @@ export type CreateUploadController = (deps: {
 export interface UploadController {
 	uploadRequest: (input: UploadInput) => Promise<UploadOutput>;
 	ackRequest: (input: AckInput) => Promise<AckOutput>;
-}
-
-export interface FileStorage {
-	getPresignUploadUrl: (
-		bucket: string,
-		objectPath: string,
-		expiration: number
-	) => Promise<string>;
-	getPresignDownloadUrl: (
-		bucket: string,
-		objectPath: string,
-		expiration: number
-	) => Promise<string>;
-}
-
-export interface SlugStorage {
-	createSlug: (input: SlugInput) => Promise<string>;
-	getValue: (slug: string) => Promise<SlugOutput | undefined>;
-	setExpiration: (slug: string, expireAt: number) => Promise<void>;
-}
-
-export interface UploadInput {
-	size: number;
-	filename: string;
-}
-
-export interface UploadOutput {
-	slug: string;
-	url: string;
-}
-
-export interface AckInput {
-	slug: string;
-}
-
-export interface AckOutput {
-	url: string;
-	expireAt: Date;
 }
